% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raindrop.R
\name{clik}
\alias{clik}
\title{clik: conditional likelihood function for a 2x2 table}
\usage{
clik(theta, aa, bb, cc, dd)
}
\arguments{
\item{theta}{log odds ratio}

\item{aa}{number in top left}

\item{bb}{number in top right}

\item{cc}{number in bottom left}

\item{dd}{number in bottom right}
}
\description{
conditional likelihood function for a 2x2 table
}
\author{
Nick Barrowman <nbarrowman@cheo.on.ca>
}
