% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raindrop.R
\name{plotraindrop}
\alias{plotraindrop}
\title{plotraindrop: plot a raindrop}
\usage{
plotraindrop(loglik, theta, y, cutoff = -2, type = "loglik",
  tfactor = 1, col = 2, lwd = 1, rot = FALSE)
}
\arguments{
\item{loglik}{Log likelihod}

\item{theta}{Corresponding parameter values.}

\item{y}{y-value}

\item{cutoff}{decrease from maximum in the log likelihood}

\item{type}{type of curve}

\item{tfactor}{not sure}
}
\description{
For plotting raindrops
}
\author{
Nick Barrowman <nbarrowman@cheo.on.ca>
}
