% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raindrop.R
\name{plotraindrops}
\alias{plotraindrops}
\title{plotraindrops: plot a list of raindrops}
\usage{
plotraindrops(logliks, theta, cutoff = -2, type = "loglik",
  tfactor = 1, col = 2, lwd = 1, xlim = range(theta),
  K = length(logliks), ylim = c(1, 3 * K + 2), xaxis = TRUE)
}
\arguments{
\item{logliks}{Log likelihoods}

\item{theta}{Corresponding parameter values.}

\item{cutoff}{decrease from maximum in the log likelihood}

\item{type}{type of curve}

\item{tfactor}{not sure}

\item{col}{fill color}

\item{lwd}{line width}

\item{xlim}{x-limits for plot}

\item{K}{number of log likelihoods (not necessary to specify)}

\item{ylim}{y-limits for plot}

\item{xaxis}{draw the x-axis?}
}
\description{
For plotting raindrops
}
\examples{
studies <- data.frame(trial=c(1,1),x=c(1,15),n=c(11,25),rx=c(1,0))
theta <- seq(-8,1,length=200)
# create quadratic log likelihood
study1fit <- glm(cbind(x,n-x)~rx,family=binomial,data=studies[studies$trial==1,])
table <- summary(study1fit)$coef
mlelogOR <- table["rx","Estimate"] 
ASElogOR <- table["rx","Std. Error"]
quadloglik <- -0.5*(theta-mlelogOR)^2/ASElogOR^2
# create conditional log likelihood
condloglik <- log(clik(theta,1,10,15,10))
maxcondloglik <- max(condloglik)
plotraindrops(list("Wald"=quadloglik,"Conditional"=condloglik),theta)

}
\author{
Nick Barrowman <nbarrowman@cheo.on.ca>
}
